Use [black](https://github.com/psf/black) to reformat all the code.
This simply makes the whole code look the same.

Note that the default is to have an 88 characters line length, so reconfigure flake8 and pycodestyle to match as per docs: https://black.readthedocs.io/en/stable/the_black_code_style.html#line-length

Note that black is not available on Python 3.5, so don't add it to the `dev` extra.