Add test cases for pcapng write support
To facilitate testing, added `__eq__` methods to `Block` and `Options`.

In doing my testing I ran into a few issues with decoding, so I've
rolled the functionality of `Block._decode()` right into
`Block.__init__()`. This way we don't need to test if we've already
decoded the block, every time we want to get a property of it.

Discovered that parsing Name Resolution Blocks added a blank entry to
the end of the list of names, so that's been fixed too.

Added/updated `.coveragerc`, but haven't (yet?) used its results to
improve our test coverage.