Completely unpack PEN from custom options
The problem is that struct.unpack returns a tuple,
regardless of whether or not the unpacked thing is a single element.

This caused a strange inconsistency in reading vs writing
a custom options:

Writing would be (0001, "Some Custom Data")
but reading would give you ((0001,), "Some Custom Data")

Additionally, this meant that read data
could not correctly be written without intervention.